requirements.txt
zenpy>=2.0.56
openai>=1.63.0
anthropic>=0.49.0
pymongo>=4.11.3
motor>=3.7.0
schedule>=1.2.0
flask>=3.1.0
requests>=2.32.3
python-dotenv>=1.0.0
ipaddress>=1.0.23
python-dateutil>=2.9.0
cachetools>=5.5.0
tqdm>=4.67.0

# Testing dependencies
pytest>=8.3.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-timeout>=2.3.0
pytest-asyncio>=0.25.0
pre-commit>=3.3.2

# Optional performance testing dependencies 
# Only install these if needed and your system supports C++ compilation
# psutil>=5.9.5
# pytest-benchmark>=4.0.0
# memory-profiler>=0.61.0
